remove potentially dangerous cast
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_3x@1190033 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/index/TermInfosReader.java
@@ -113,7 +113,7 @@ final class TermInfosReader implements Closeable {
         final SegmentTermEnum indexEnum = new SegmentTermEnum(directory.openInput(indexFileName,
                                                                                   readBufferSize), fieldInfos, true);
         try {
-          index = new TermInfosReaderIndex(indexEnum, indexDivisor, (int) dir.fileLength(indexFileName), totalIndexInterval);
+          index = new TermInfosReaderIndex(indexEnum, indexDivisor, dir.fileLength(indexFileName), totalIndexInterval);
           indexLength = index.length();
         } finally {
           indexEnum.close();
